+++ core/src/test/scala/com/microsoft/azure/synapse/ml/nbtest/DatabricksTests.scala
@@ -3,53 +3,18 @@
 
 package com.microsoft.azure.synapse.ml.nbtest
 
-import com.microsoft.azure.synapse.ml.core.test.base.TestBase
 import com.microsoft.azure.synapse.ml.nbtest.DatabricksUtilities._
 
 import java.util.concurrent.TimeUnit
-import scala.collection.mutable
+import scala.collection.mutable.ListBuffer
 import scala.concurrent.Await
 import scala.concurrent.duration.Duration
 import scala.language.existentials
 
-class DatabricksTests extends TestBase {
+class DatabricksTests extends DatabricksTestProcess {
 
   val clusterId: String = createClusterInPool(ClusterName, AdbRuntime, PoolId)
-  val jobIdsToCancel: mutable.ListBuffer[Int] = mutable.ListBuffer[Int]()
-
-  println("Checking if cluster is active")
-  tryWithRetries(Seq.fill(60 * 15)(1000).toArray) { () =>
-    assert(isClusterActive(clusterId))
-  }
-  println("Installing libraries")
-  installLibraries(clusterId, Libraries)
-  tryWithRetries(Seq.fill(60 * 3)(1000).toArray) { () =>
-    assert(areLibrariesInstalled(clusterId))
-  }
-  println(s"Creating folder $Folder")
-  workspaceMkDir(Folder)
-
-  println(s"Submitting jobs")
-  val parNotebookRuns: Seq[DatabricksNotebookRun] = ParallelizableNotebooks.map(uploadAndSubmitNotebook(clusterId, _))
-  parNotebookRuns.foreach(notebookRun => jobIdsToCancel.append(notebookRun.runId))
-
-  println(s"Submitted ${parNotebookRuns.length} for execution: ${parNotebookRuns.map(_.runId).toList}")
-
-  assert(parNotebookRuns.nonEmpty)
-
-  parNotebookRuns.foreach(run => {
-    println(s"Testing ${run.notebookName}")
-
-    test(run.notebookName) {
-      val result = Await.ready(
-        run.monitor(logLevel = 0),
-        Duration(TimeoutInMillis.toLong, TimeUnit.MILLISECONDS)).value.get
-
-      if (!result.isSuccess){
-        throw result.failed.get
-      }
-    }
-  })
+  val jobIdsToCancel: ListBuffer[Int] = databricksTestProcess(clusterId, Libraries, ParallelizableNotebooks)
 
   println(s"Submitting nonparallelizable job...")
   NonParallelizableNotebooks.toIterator.foreach(notebook => {
@@ -68,11 +33,7 @@ class DatabricksTests extends TestBase {
   })
 
   protected override def afterAll(): Unit = {
-    println("Suite DatabricksTests finished. Running afterAll procedure...")
-    jobIdsToCancel.foreach(cancelRun)
-
-    deleteCluster(clusterId)
-    println(s"Deleted cluster with Id $clusterId.")
+    afterAllHelper(jobIdsToCancel, clusterId)
 
     super.afterAll()
   }
